requirements.txt
fastapi==0.116.1
sqlalchemy==2.0.43
asyncpg==0.30.0
python-dotenv==1.1.1
pytest==8.4.1
httpx==0.28.1
uvicorn[standard]==0.35.0
